# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VisionHub360",
        "description": "API RESTful con JWT y Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/documentsClient": {
            "post": {
                "tags": [
                    "Gestión de Documentos de Clientes"
                ],
                "summary": "Registrar un nuevo documento de cliente con múltiples imágenes",
                "operationId": "6ecdd9ccfdd8625e14608894337ba3ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "name_company"
                                ],
                                "properties": {
                                    "name_company": {
                                        "type": "string",
                                        "example": "Joyas del Sur S.A."
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Documentos de garantía del cliente"
                                    },
                                    "client_id": {
                                        "type": "integer",
                                        "example": 56
                                    },
                                    "document_files[]": {
                                        "description": "Múltiples imágenes del documento",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Documento de cliente creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "403": {
                        "description": "Cliente no pertenece a la empresa"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/documentsClient/{client_id}/history": {
            "get": {
                "tags": [
                    "Gestión de Documentos de Clientes"
                ],
                "summary": "Listar historial de interacciones de documentos del cliente con sus imágenes y sucursal",
                "operationId": "e6d0cc07e6689a821c67bf150a37e9aa",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 56
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historial de documentos del cliente obtenido correctamente"
                    },
                    "403": {
                        "description": "Cliente no pertenece a la empresa"
                    },
                    "404": {
                        "description": "El cliente no tiene documentos registrados"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/document-clients/{document_id}/image/{image_id}": {
            "delete": {
                "tags": [
                    "Gestión de Documentos de Clientes"
                ],
                "summary": "Eliminar una imagen específica de un documento de cliente",
                "operationId": "9af44c49127da18c9bd4e89ef7907074",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "description": "ID del documento principal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 25
                        }
                    },
                    {
                        "name": "image_id",
                        "in": "path",
                        "description": "ID de la imagen a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 53
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagen eliminada y documento actualizado correctamente"
                    },
                    "403": {
                        "description": "Acceso denegado o documento/imagen no pertenecen a la empresa"
                    },
                    "404": {
                        "description": "El documento o la imagen no fueron encontrados"
                    },
                    "500": {
                        "description": "Error interno del servidor al eliminar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/document-clients/{document_id}": {
            "post": {
                "tags": [
                    "Gestión de Documentos de Clientes"
                ],
                "summary": "Actualizar datos y/o agregar imágenes a una interacción (documento) existente",
                "operationId": "835745591a83c15eb1f8bd7de748bb86",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "description": "ID del documento principal a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 27
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name_company": {
                                        "description": "Nombre de la compañía (Opcional)",
                                        "type": "string",
                                        "example": "Nueva S.A.S."
                                    },
                                    "description": {
                                        "description": "Descripción del documento (Opcional)",
                                        "type": "string",
                                        "example": "Nuevos documentos de garantía."
                                    },
                                    "images[]": {
                                        "description": "Array de nuevos archivos de imagen a subir (Opcional)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Documento e imágenes actualizados correctamente"
                    },
                    "403": {
                        "description": "Acceso denegado o documento no pertenece a la empresa"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Documento no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}